easy way out..
moved away from trying to find a solution for generic grouping of nested list.

This means that this solution cannot solve for problems that have more than two duplicates on the same row.

for example if one row has 3 duplicates, the solution will not be corret.

This is an issue that i indent to revisit given enough time